formats: Disable ASTC on GL hosts
For GL hosts CopyImageSubData and TextureView are not defined
when ASTC textures are involved, and the copy fallback currently
doesn't handle these formats correctly. Since Gallium can handles
ASTC texures transparetly in the guest, host support is not really
required to be able to use it.

Fixes all tests out of
  dEQP-GLES32.functional.copy_image*astc*
that were failing on the llvmpipe GL host before

Related #224

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>