shader: Link shader command
Add a new command associated to glLinkProgram. With this we should be
able to compile and link shaders when requested by the user.

Together with the command we expect an array of shader handles attached
to the program, where each position of the array corresponds to a pipe
shader type.

Signed-off-by: Antonio Caggiano <antonio.caggiano@collabora.com>
Reviewed-by: Gert Wollny <gert.wollny@collabora.com>