virgl: add VIRGL_RENDERER_RENDER_SERVER
When the flag is set, a render server is started and each vkr context is
replaced by a proxy context.

It does not apply to vrend contexts.  There is no plan to support vrend
contexts either currently.

This also extends virgl_renderer_callbacks with a new callback.  The
idea is to allow the client to start the render server externally.  It
can be used to sandbox the render server, or is required when the client
process is sandboxed and cannot fork/exec/socketpair.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Reviewed-by: Ryan Neph <ryanneph@google.com>