Make COPY_TRANSFER3D bidirectional
COPY_TRANSFER3D was designed to handle transfers
to host. This patch is making it possible to use
this CCMD in both directions.

If guest would like to have this data available
immediately, then it should encode, flush and wait on
the guest side.

The data is copied from host to iovec attached to resource.
In most cases resource is a staging buffer.

Reviewed-by: Gert Wollny <gert.wollny@collabora.com>