vkr: add support for blob_id 0
Return a memfd/shm when a blob resource of blob id 0 is exported.  This
becomes a part of the venus protocol.

This aims to replace the use of VIRGL_RENDERER_BLOB_MEM_GUEST in venus.
There are two reasons.  One is that VIRGL_RENDERER_BLOB_MEM_GUEST cannot
be shared and does not work when we add a render server (unless a
udmabuf is created from iov).  The other is that it gives us a logically
contiguous mapping.  We can stop dealing with iov.

Another option is achieve the same purposes is to introduce
VIRGL_RENDERER_BLOB_MEM_HOST2D.  That is more generic, but requires
changes to virtio-gpu spec and to the kernel.

Yet another option is to add a new Vulkan object, VkShmemMESA, and
commands to create/destroy the object.  Then we can use a non-zero blob
id to export to the Vulkan object.  But that feels like an overkill
because all we need to allocate a shmem is its size, which is already
given.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Reviewed-by: Ryan Neph <ryanneph@google.com>