ci: Make virtio-gpu test results available in the artifacts
Because the tests run using the Mesa CI infrastructure, the results
folder is created in a Mesa project specific location and it is
necessary to move it to CI_PROJECT_DIR once completed.

This operation has been performed only for the vtest based jobs, hence
let's ensure it is always executed.

Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>
Reviewed-by: Guilherme Gallo <guilherme.gallo@collabora.com>
Acked-by: Daniel Stone <daniels@collabora.com>