drm: Add msm native context implementation
The idea is to add new context type(s) which allows using native Mesa
drivers in a guest userspace, talking their own protocol to a small host
side component which interfaces with the native host drm driver.  The
usermode driver in the guest uses the virtio_gpu driver as the "transport
layer" for that protocol, as well as for handling GEM buffer mapping
(and dma-buf/dma-fence sharing between host and guest, etc).  If the
protocol is designed properly to avoid synchronous round trips between
host and guest in hot paths, the performance can match native (outside
of VM), which can be a significant improvement over API virtualization.
(Especially on lower powered devices where the overhead of running a gl
driver on top of a gl driver is prohibitive.)

This patch adds a new context type to allow the guest to use mesa/freedreno
on top of virtgpu "passing through" the drm/msm interface to the guest.

The interface does not map 1:1 to host kernel interface in all cases, to
reduce the number of round trips.  There could be further possible
improvements with the addition of some new kernel UABI, for example if
we allowed userspace to allocate GPU virtual addresses, then allocation
of new GEM buffers could be made asynchronous.

Signed-off-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Dmitry Osipenko <dmitry.osipenko@collabora.com>
Reviewed-by: Ryan Neph <ryanneph@google.com>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Reviewed-by: Chia-I Wu <olvaffe@gmail.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/693>