ci: Optimize virtio-gpu jobs execution
Since crosvm is not able to properly handle parallel tests due to a
concurrency bottleneck, it's necessary to ensuring scaling at the host
system level by increasing the number of active crosvm instances.

For dEQP testing this can be easily accomplished by using the Deqp
Suite support in deqp-runner and adding the crosvm wrapper script in
the related '*.toml' files.

For piglit this is currently not supported in piglit-runner, but
something similar could be added later. For the moment we use an
increased number of parallel jobs to speed-up the execution.

Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>
Reviewed-by: Guilherme Gallo <guilherme.gallo@collabora.com>
Acked-by: Daniel Stone <daniels@collabora.com>