vkr: Create vkr_allocator
A Vulkan allocator for allocating and mapping `virgl_resource`s
via opaque fds.

A dedicated host context is created to store info and
handles pertaining to the host device.

When mapping a `virgl_resource`, the host context will import
the VkDeviceMemory and map it directly to the provided host
pointer. The device is selected by comparing the device UUID
from which the `virgl_resource` was created.

Device UUID is the specification's answer to ensuring that
devices can be correlated across API and driver boundaries.
After mapping, the parameters necessary to free and unmap the
memory are stored in a `vkr_opaque_fd_mem_info` struct and
tracked by the host context.

Signed-off-by: Omar Akkila <omar.akkila@collabora.com>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/651>