shader: Make sure that we are not mixing generic and patch variables
It is possible for two variable to have the same sid and array_id as long as they
have a different name. Make sure that we are not considering them as two components.

Also fix the already_found_one variable define in the loop when we need it to persist
between iterations.

Signed-off-by: Corentin Noël <corentin.noel@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/736>