vrend: synchronize fence completion and query checking when using async_cb
When using the async fence callback, there is no longer a need for
calling code to periodically poll the renderer for fence completion, but
vrend still requires it to avoid issues with query checking.

Now, when use_async_cb is true, fences are completed via callback in the
sync thread. When pending queries exist, they must first be checked on
the main thread before each fence can be signaled to the caller. This is
done by signaling an eventfd that the caller must react to by waking and
calling virgl_renderer_poll() on the main thread. After query checking,
the sync thread is woken to signal the fence.

When no pending queries exist, there is no need to signal the eventfd,
and fences are completed immediately by the sync thread without the
additional blocking.

Signed-off-by: Ryan Neph <ryanneph@google.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/700>