ci: Don't run post-merge test jobs
We mark the build job as never running post-merge, because we just don't
want a pipeline running. GitLab seems to respond to this by just trying
to run the job anyway, so mark it as never.

In all other cases, it will be run on_success when the upstream jobs it
depends on have completed.

Signed-off-by: Daniel Stone <daniels@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/748>