shader: Take the output range into account when storing the so_names
Make sure to take the generic/patch output range into consideration when
storing the so_names as we need the actual name to be taken into account.

This fixes the `spec@ext_transform_feedback@max-varyings` Piglit test with the
NIR-to-TGSI branch.

Signed-off-by: Corentin Noël <corentin.noel@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/776>