drm/msm: Drop reloc dumping
Relocs are not supported by msm renderer.  But since the cmds contain
guest controlled data, nr_relocs could be non-zero and relocs pointer
pointing to a bogus address.

We already validate the bos and cmds tables are of an appropriate size,
so it is safe to continue dumping them.

Also, limit this to debug builds, since it is stricty a debugging
feature.

Signed-off-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Chia-I Wu <olvaffe@gmail.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/795>