vrend: always reemit GL_DEPTH_TEXTURE_MODE
The reason this condition exists is to avoid an extra call to
glTexParameteri when the depth texture mode changes. But this is hardly
ever true, and even when it is, we already call glTexParameteri multiple
times anyway, so an extra call here doesn't really matter and just makes
the code slightly more complex by adding more state to keep track of.

Signed-off-by: Italo Nicola <italonicola@collabora.com>
Reviewed-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/792>