vrend: Add flag that will advertise SSO support to the guest
We did a mistake in the guest that the TGSI_SEPARABLE_SHADER
property would be send without the host actually understanding
this proeprty, so add a flag that can signal this support.

Hold off in setting it until we have some actual code to deal
with the property and possibly with additional information.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/846>