shader: pass info about written outputs and create inputs when missing
With NTT we don't get any reasonable information about arrays anymore, and
therefore we have to re-created the arrays when needed. However, it is legal
not to consume all outputs, and NIR will optimize away unused inputs,
unless a shader is compiled as separable shader. Therefore, if we don't
have separable shaders we have to re-create the inputs to be able to
create the corresponding arrays and correct access.

v2: Fix typo (Corentin)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/811>