vrend: implement polygon stipple by using the system values UBO
Using a texture to emulate the polygon stipple requires a texture and
with that a uniform slot in the default UBO. Thus, in order to be exact
and account for the possibility that a guest application want's to use
polygon stipples, we had to substract "one" from the number of available
uniform values in FS. This didn't play well with Tomb Raider 2013 because
it results in dropping the number of available uniforms below the number
of uniforms required by OpenGL 4.3.

Since the stipple pattern is just a number of bits that need testing it
this emulation can also be implemented by storing the pattern in the
system values UBO that is already in use. With that we can drop the
reservation of the extra uniform in the FS.

v2: - use a define for the stipple array size and a 2D array
      for the pattern (Corentin)
    - add a static assert and make it clear where the mask comes from
      when emitting the shader code

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/857>