vrend: work around glamor rejecting llvmpipe
glamor checks in the guest whether the render string
has "llvmpipe" in it, and if this is the case, then it
is not loaded. The result is that glx/egl don't pick up
virgl as a renderer and we only get llvmpipe as OpenGL
driver.

Work around this by replacing "llvmpipe" with the
uppercase version.

v2: uses memcpy to silence warning
    (warning pointed out by Dmitry)

v3: Fix CI renderer check (Corentin)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Corentin Noël <corentin.noel@collabora.com>
Tested-by: Dmitry Osipenko <dmitry.osipenko@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/865>