shader: Stabilize the location assignment
While we can always assume that the number of varyings is 32, the
number of patch varyings might be lower, so instead of counting
down for patches and counting up for generics, do the opposite.

In addition, use a better formatting of the emitted glsl around
this declaration. Since we no longer patch the IO variables,
there is no need to format this to some specific pattern.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Italo Nicola <italonicola@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/871>