TGSI: don't reject the image index -1 right away
With the change to NTT the guest might now send an image index
of -1, e.g. with
 spec@arb_shader_image_load_store@execution@image-array-out-of-bounds-access-load

Instead of bailing out with an error, and thereby leaving the
host in an error state, accept the negative index, and let the
shader translation handle it.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Ryan Neph <ryanneph@google.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/896>