shader: Work around out-of-bounds buffer and image access
A negative index is an obvious indicator for an out-of bound access, and
for images we also test whether the image index is among the used images
to check whether the access is not OOB.

In case of an OOB access replace a load with assigning zero to the
destination register, and drop the write instruction altogether.

Making TGSI accpet a negative index emitted by the guest, will be
added with the next commit.

v2: add assertions (Ryan)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Ryan Neph <ryanneph@google.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/896>