vkr: Re-enable VK_EXT_private_data
Even though we implement the extension natively in the guest driver, as
a workaround we still require support in the host driver too. The guest
driver does not yet scrub this extension's stuff from the pNext chain of
VkDeviceCreateInfo and VkPhysicalDeviceFeatures2, so guest usage of the
extension will trigger invalid usage in the host if the host does not
support it.

See: https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/18000
Signed-off-by: Chad Versace <chadversary@chromium.org>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/908>