vrend: glColorMask updates all color masks, act accordingly
When glColorMask is called, all color masks are updated to the
given mask, so also update the masks stored in the state to
have the value of the mask given for draw buffer 0.

This fixes a state leak where color masks for independent blend
would not be updated correctly in all cases.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/954>