server: rename render_virgl to render_state
This is to prepare for virgl decouple. We need the state tracking only
for threaded render worker. Removing the thread config makes the
forwarding much cleaner but reduces debuggability. So I prefer to keep
this shim layer but with the new naming.

Signed-off-by: Yiwei Zhang <zzyiwei@chromium.org>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/993>