ci: Uprev mesa and fix the debian/x86_test-{gl,vk} references
Mesa has been updated because of confusion in some variables to define the
test images when one likes to use GL or VK. Because of that, the uprev is
necessary together with a change in naming those variables. So, the jobs
extending from ".use-{gl,vk}-test-image" will now be sure they are using the
correct image.

Signed-off-by: Sergi Blanch Torne <sergi.blanch.torne@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/987>