virgl: introduce virgl_renderer_resource_get_info_ext()
It's becoming commonplace that DRM drivers make use of 'modifiers'
that alter the in-memory representation of a texture, which means that
when importing a texture the same exact 'modifiers' must be specified
along with the other usual parameters.

Introduce a new structure, virgl_renderer_resource_info_ext, that
extends upon the original virgl_renderer_resource_info with the
'planes' and 'modifiers' fields, and add a new function,
virgl_renderer_resource_get_info_ext() that performs the same actions
as virgl_renderer_resource_get_info() but using the new extended
structure.

Signed-off-by: Sergio Lopez <slp@redhat.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1024>