vrend: Create a texture view if the image view doesn't cover the whole array
When an image is bound we can only bind just one layer or all layers. If the
guest created a texture view of the image to bind a sub-set of the layers, then
we have to do the same.

Fixes piglit: arb_texture_view-rendering-layers-image

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1060>