vtest: make client disconnect more intuitive
When the app destroys the vtest renderer, the vtest client shuts down
the socket or directly close it, and the detection of such on the server
side is via a read with zero bytes returned without errors. This change
adds a non-error result code for such case to avoid confusions.

Signed-off-by: Yiwei Zhang <zzyiwei@chromium.org>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1076>