shader/tgsi: check array access and report failure when needed
Originally in mesa TGSI was never intended to be interfaced from outside
the library so that one could rely on the code to adhere the limits for
inputs, outputs etc. However, with virglrenderer we receive TGSI as text
and the text parser doesn't check the limits, so that the actual TGSI
might be illegal and result in OOB access.

Since checking these limits when parsing the text would be more tedious,
check these limits in the tgsi info parser that is always run before when
the TGSI is translated and report errors.

Closes: https://gitlab.freedesktop.org/virgl/virglrenderer/-/issues/339

v2: fix ws (Corentin)

Signed-off-by: Gert Wollny <gert.wollny@collabora.co.uk>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1105>